refactor[#403] : ThumbnailList 재사용성 적어 ThumbnailGrid 컴포넌트 내부에서 정의
