+++ src/entities/marking/ui/index.ts
@@ -1,3 +1,2 @@
 export * from "./temporaryMarkingBar";
 export * from "./dashboardMarkingThumbnail";
-export * from "./thumbnailList";
